Object creation now lives in the ApplicationFactory.  Trying to be as true to the intended purpose of this pattern as possible.  It does mean that the AF has some responsibilities outside of object creation (i.e. kicking off the display of the UI).  I might change this in the future, but for now it seems pretty logical.
Having stuff going out and looking in the AF for stuff (with getters) was probably a bad idea.  It's better that stuff declares its dependencies in the constructor.  That's what I've started doing.  I realise that in the old Jdig I didn't really do this, partially out of laziness, partially out of the sheer complexity that arises out of using Swing and the insane-o amount of references that ended up flying around.

We'll see how this goes.